% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeVCF_haplo.R
\name{imputeVCF_haplo}
\alias{imputeVCF_haplo}
\title{Impute Inbred VCF per defined window}
\usage{
imputeVCF_haplo(P_in_vcf = "", IP_in_vcf = "", IP_new_out_vcf = "", win = 5)
}
\arguments{
\item{P_in_vcf}{Parents original VCF}

\item{IP_in_vcf}{Inbred Populations original VCF}

\item{IP_new_out_vcf}{Name for an output - Imputed Inbred VCF}

\item{win}{Size of the discrete window}
}
\value{
Imputed inbred populations VCF
}
\description{
This function impute VCF data from the inbred populations based on the VCF data from the parents
}
