% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeVCF_singlesnp.R
\name{imputeVCF_singlesnp}
\alias{imputeVCF_singlesnp}
\title{Impute Offspring VCF per SNP}
\usage{
imputeVCF_singlesnp(P_in_vcf = "", OS_in_vcf = "", OS_new_out_vcf = "")
}
\arguments{
\item{P_in_vcf}{Parents original VCF}

\item{OS_in_vcf}{Offspring original VCF}

\item{OS_new_out_vcf}{Name for an output - Imputed Offspring VCF}
}
\value{
Imputed Offspring VCF
}
\description{
This function impute VCF data from the offsprings based on the VCF data from the parents
}
