% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspringsim.R
\name{offspringsim}
\alias{offspringsim}
\title{Simulate Offspring Genotpyes}
\usage{
offspringsim(parentdf = parentdf, prop = prop, nsamples = nsamples)
}
\arguments{
\item{parentdf}{Data frame containing the parents genotype - generated from parentsim function}

\item{prop}{Proportion of original number os SNPs retained in the final simulated genotype - subset function}

\item{nsamples}{Number of offsprings}
}
\value{
A data frame with the simulated genotypes and genomic information for the offsprings
}
\description{
This function simulate a genotype data for one chromosome for a given number of samples based on parents genotypes
}
