% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlesnpinp.R
\name{singlesnpinp}
\alias{singlesnpinp}
\title{Impute the Genotypes per SNP}
\usage{
singlesnpinp(OS = OS, P = P)
}
\arguments{
\item{OS}{Offspring genotypic data}

\item{P}{Parents genotypic data}
}
\value{
Imputed Offspring Data
}
\description{
This function impute data in the offspring based on the parents genotypes
}
